Move snackbar stuff into own folder
