+++ ui/src/stores/WebSocketStore.ts
@@ -1,4 +1,4 @@
-import {SnackReporter} from './SnackManager';
+import {SnackReporter} from '../snack/SnackManager';
 import {CurrentUser} from './CurrentUser';
 import * as config from '../config';
 import {MessagesStore} from './MessagesStore';
